move program/src/wasm/hash.rs contents to solana-hash crate
