+++ sdk/program/src/wasm/mod.rs
@@ -2,7 +2,6 @@
 #![cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
 
-pub mod hash;
 pub mod instructions;
 pub mod pubkey;
 pub mod system_instruction;
